--- infra/bridge/src/hooks/index.js
@@ -1,11 +1,0 @@
-'use strict'
-
-const { subscribeToHooks: subscribeToTwitterHooks } = require('./twitter')
-
-async function subscribeToHooks() {
-  return subscribeToTwitterHooks()
-}
-
-module.exports = {
-  subscribeToHooks
-}
